+++ docs/src/examples/components/Radio/States/RadioExampleChecked.shorthand.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Radio } from '@stardust-ui/react'
+
+const RadioExampleCheckedShorthand = () => (
+  <Radio label="This radio comes pre-checked" defaultChecked />
+)
+
+export default RadioExampleCheckedShorthand