+++ docs/src/examples/components/Radio/index.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react'
 import Types from './Types'
+import States from './States'
 
 const RadioExamples = () => (
   <div>
     <Types />
+    <States />
   </div>
 )